Added basic testing to test testing
