+++ gamersnet_backend/README.md
@@ -7,4 +7,5 @@ npm install
 ```
 
 ## Running
-To run locally execute `npm start`.
+To run locally execute `npm start`.  
+To test, include any test files in a folder named `__tests__` and execute `npm test`.
